refactor ReportFilter to component
