+++ app/routes/reports.$filters.tsx
@@ -1,0 +1,3 @@
+export default function Filters() {
+	return <div>filtered reports go here</div>;
+}
